+++ boofuzz/__init__.py
@@ -10,6 +10,7 @@ from . import blocks, exception, legos, primitives
 from .blocks import Block, Checksum, Repeat, Request, REQUESTS, Size
 from .connections import (
     BaseSocketConnection,
+    FileConnection,
     ip_constants,
     ISerialLike,
     ITargetConnection,
@@ -24,7 +25,7 @@ from .connections import (
 )
 from .constants import BIG_ENDIAN, DEFAULT_PROCMON_PORT, LITTLE_ENDIAN
 from .event_hook import EventHook
-from .exception import MustImplementException, SizerNotUtilizedError, SullyRuntimeError
+from .exception import MustImplementException, SizerNotUtilizedError, SullyRuntimeError, BoofuzzFailure
 from .fuzz_logger import FuzzLogger
 from .fuzz_logger_csv import FuzzLoggerCsv
 from .fuzz_logger_curses import FuzzLoggerCurses
@@ -67,6 +68,7 @@ __all__ = [
     "BitField",
     "Block",
     "blocks",
+    "BoofuzzFailure",
     "Byte",
     "Bytes",
     "CallbackMonitor",
@@ -77,6 +79,7 @@ __all__ = [
     "DWord",
     "EventHook",
     "exception",
+    "FileConnection",
     "FromFile",
     "FuzzLogger",
     "FuzzLoggerCsv",