+++ boofuzz/exception.py
@@ -1,6 +1,19 @@
 import attr
 
 
+@attr.s
+class BoofuzzFailure(Exception):
+    """Raises a failure in the current test case.
+
+    This is most important when failures in the target may not be noticed until the next test case. In such a situation,
+    it may be necessary to abort the test case before a fuzz message is even sent, as the fuzz message may be poorly
+    defined outside the context of a valid partial protocol exchange.
+    """
+
+    message = attr.ib(type=str, default="")
+    pass
+
+
 class BoofuzzError(Exception):
     pass