+++ boofuzz/fuzz_logger_curses.py
@@ -8,9 +8,9 @@ try:
     import curses  # pytype: disable=import-error
 except ImportError:
     # Allow fuzz_logger_curses to be imported on Windows -- will fail if you try to use it.
-    warnings.warn("Importing curses failed. Console GUI features will not be available.", UserWarning, stacklevel=2)
+    warnings.warn("Importing curses failed. Optional console GUI features will not be available.", UserWarning)
     curses = None
-    pass
+
 import signal
 import threading
 
@@ -29,11 +29,19 @@ else:
         def get_terminal_size():
             return [130, 40]
 
-        warnings.warn(
-            "Console GUI will not resize properly. Install shutil_backports for full support.",
-            UserWarning,
-            stacklevel=2,
-        )
+        warnings.warn("Console GUI will not resize properly. Install shutil_backports for full support.", UserWarning)
+
+COLOR_PAIR_WHITE = 1
+COLOR_PAIR_CYAN = 2
+COLOR_PAIR_RED = 3
+COLOR_PAIR_YELLOW = 4
+COLOR_PAIR_GREEN = 5
+COLOR_PAIR_MAGENTA = 6
+COLOR_PAIR_BLACK = 7
+
+STATUS_RUNNING = 0
+STATUS_PAUSED = 1
+STATUS_DONE = 2
 
 
 class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
@@ -41,7 +49,6 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
     This class formats FuzzLogger data for a console GUI using curses. This hasn't been tested on Windows.
     """
 
-    DEFAULT_HEX_TO_STR = helpers.hex_to_hexstr
     INDENT_SIZE = 2
 
     def __init__(
@@ -49,25 +56,25 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
         web_port=26000,
         window_height=40,
         window_width=130,
-        auto_scoll=True,
+        auto_scroll=True,
         max_log_lines=500,
         wait_on_quit=True,
         min_refresh_rate=1000,
-        bytes_to_str=DEFAULT_HEX_TO_STR,
+        bytes_to_str=helpers.hex_to_hexstr,
     ):
         """
         :type web_port: int
         :param web_port: Webinterface port. Default 26000
 
         :type window_height: int
-        :param window_height: Default console heigth, set to on startup. Default 40
+        :param window_height: Default console height, set to on startup. Default 40
 
         :type window_width: int
         :param window_width: Default console width, set to on startup. Default 130
 
-        :type auto_scoll: bool
-        :param auto_scoll: Whether to auto-scoll the cases and crashed windows to allways display the last line if there
-                           are too many lines to display all of them. Default True
+        :type auto_scroll: bool
+        :param auto_scroll: Whether to auto-scroll the cases and crashed windows to always display the last line if
+                            there are too many lines to display all of them. Default True
 
         :type max_log_lines: int
         :param max_log_lines: Maximum log lines to keep in the internal storage. Additional lines exceeding this limit
@@ -88,13 +95,13 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
         self._title = "boofuzz"
         self._web_port = web_port
         self._max_log_lines = max_log_lines
-        self._auto_scroll = auto_scoll
+        self._auto_scroll = auto_scroll
         self._current_data = None
         self._log_storage = []
         self._fail_storage = []
         self._wait_on_quit = wait_on_quit
         self._quit = False
-        self._status = 0  # 0: Running 1: Paused 2: Done
+        self._status = STATUS_RUNNING
         self._refresh_interval = min_refresh_rate
         self._event_resize = True
         self._event_log = False
@@ -124,16 +131,16 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
         curses.use_default_colors()
         curses.noecho()
         curses.curs_set(0)
-        self._stdscr.nodelay(1)
+        self._stdscr.nodelay(True)
 
         # Curses color pairs
-        curses.init_pair(1, curses.COLOR_WHITE, -1)
-        curses.init_pair(2, curses.COLOR_CYAN, -1)
-        curses.init_pair(3, curses.COLOR_RED, -1)
-        curses.init_pair(4, curses.COLOR_YELLOW, -1)
-        curses.init_pair(5, curses.COLOR_GREEN, -1)
-        curses.init_pair(6, curses.COLOR_MAGENTA, -1)
-        curses.init_pair(7, curses.COLOR_BLACK, curses.COLOR_WHITE)
+        curses.init_pair(COLOR_PAIR_WHITE, curses.COLOR_WHITE, -1)
+        curses.init_pair(COLOR_PAIR_CYAN, curses.COLOR_CYAN, -1)
+        curses.init_pair(COLOR_PAIR_RED, curses.COLOR_RED, -1)
+        curses.init_pair(COLOR_PAIR_YELLOW, curses.COLOR_YELLOW, -1)
+        curses.init_pair(COLOR_PAIR_GREEN, curses.COLOR_GREEN, -1)
+        curses.init_pair(COLOR_PAIR_MAGENTA, curses.COLOR_MAGENTA, -1)
+        curses.init_pair(COLOR_PAIR_BLACK, curses.COLOR_BLACK, curses.COLOR_WHITE)
 
         # Start thread and restore the original SIGWINCH handler
         self._draw_thread = threading.Thread(name="curses_logger", target=self._draw_screen)
@@ -143,11 +150,15 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
 
     def open_test_case(self, test_case_id, name, index, *args, **kwargs):
         self._log_storage = []
-        self._total_index = index
-        self._total_num_mutations = kwargs["num_mutations"]
         self._current_name = name
-        self._current_index = kwargs["current_index"]
-        self._current_num_mutations = kwargs["current_num_mutations"]
+        self._total_index = index
+        if "current_index" in kwargs:
+            self._current_index = kwargs["current_index"]
+        if "current_num_mutations" in kwargs:
+            self._current_num_mutations = kwargs["current_num_mutations"]
+        if "num_mutations" in kwargs:
+            self._total_num_mutations = kwargs["num_mutations"]
+
         self._log_storage.append(
             helpers.format_log_msg(msg_type="test_case", description=test_case_id, format_type="curses")
         )
@@ -179,7 +190,7 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
             + (4 * indent_size + 1 - len(str(self._total_index))) * " "
             + description.strip()
         )
-        self._fail_storage.append([fail_msg, 1])
+        self._fail_storage.append([fail_msg.replace("\n", " "), COLOR_PAIR_WHITE])
         self._log_storage.append(helpers.format_log_msg(msg_type="fail", description=description, format_type="curses"))
         self._event_crash = True
         self._event_log = True
@@ -191,7 +202,7 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
             + (4 * indent_size + 1 - len(str(self._total_index))) * " "
             + description.strip()
         )
-        self._fail_storage.append([fail_msg, 3])
+        self._fail_storage.append([fail_msg.replace("\n", " "), COLOR_PAIR_RED])
         self._log_storage.append(
             helpers.format_log_msg(msg_type="error", description=description, format_type="curses")
         )
@@ -210,13 +221,16 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
         self._event_case_close = True
 
     def close_test(self):
-        self._status = 2
+        self._status = STATUS_DONE
         self._quit = True
-        self._draw_thread.join()
-        curses.nocbreak()
-        self._stdscr.keypad(False)
-        curses.echo()
-        curses.endwin()
+        try:
+            self._draw_thread.join()
+        finally:
+            self._wait_on_quit = False
+            curses.nocbreak()
+            self._stdscr.keypad(False)
+            curses.echo()
+            curses.endwin()
 
     def _draw_main(self, force=False):
         self._width, self._height = get_terminal_size()
@@ -235,45 +249,45 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
         # Render title
         self._stdscr.addstr(0, 0, "=" * self._width)
         start_x_title = int((self._width // 2) - (len(self._title) // 2) - len(self._title) % 2)
-        self._stdscr.addstr(0, start_x_title, self._title, curses.color_pair(2) | curses.A_BOLD)
+        self._stdscr.addstr(0, start_x_title, self._title, curses.color_pair(COLOR_PAIR_CYAN) | curses.A_BOLD)
 
         # Render status bar
-        self._stdscr.attron(curses.color_pair(7))
+        self._stdscr.attron(curses.color_pair(COLOR_PAIR_BLACK))
         self._stdscr.addstr(self._height - 1, 0, " " * (self._width - 1))
         self._stdscr.insch(" ")  # Fill bottom right corner
         if self._quit:
-            self._stdscr.addstr(self._height - 1, 1, "Press 'q' to quit", curses.color_pair(7) | curses.A_BLINK)
+            self._stdscr.addstr(
+                self._height - 1, 1, "Press 'q' to quit", curses.color_pair(COLOR_PAIR_BLACK) | curses.A_BLINK
+            )
         else:
             self._stdscr.addstr(self._height - 1, 1, "Press 'CTRL+C' to abort")
         self._stdscr.addstr(self._height - 1, self._width - len(self._version) - 1, self._version)
-        self._stdscr.attroff(curses.color_pair(7))
+        self._stdscr.attroff(curses.color_pair(COLOR_PAIR_BLACK))
         self._stdscr.refresh()
 
         # Initialise test case window
-        # pytype: disable=module-attr
         self._casescr_frame = curses.newpad(self._max_log_lines + 1, self._width)
-        self._casescr_frame.nodelay(1)
-        self._casescr_frame.border(0, 0, 0, " ", 0, 0, curses.ACS_VLINE, curses.ACS_VLINE)
-        self._casescr_frame.addstr(0, 1, "Test case log", curses.color_pair(4) | curses.A_BOLD)
+        self._casescr_frame.nodelay(True)
+        self._casescr_frame.border()
+        self._casescr_frame.addstr(0, 1, "Test case log", curses.color_pair(COLOR_PAIR_YELLOW) | curses.A_BOLD)
         self._casescr_frame.refresh(0, 0, 1, 0, self._height - 18, self._width)
         self._casescr = self._casescr_frame.subpad(self._max_log_lines, self._width - 2, 1, 1)
         self._draw_case()
 
         # Initialise crash window
         self._crashescr_frame = curses.newpad(self._max_log_lines + 1, self._width)
-        self._crashescr_frame.nodelay(1)
-        self._crashescr_frame.border(0, 0, 0, " ", 0, 0, curses.ACS_VLINE, curses.ACS_VLINE)
-        self._crashescr_frame.addstr(0, 1, "Crashes", curses.color_pair(3) | curses.A_BOLD)
+        self._crashescr_frame.nodelay(True)
+        self._crashescr_frame.border()
+        self._crashescr_frame.addstr(0, 1, "Crashes", curses.color_pair(COLOR_PAIR_RED) | curses.A_BOLD)
         self._crashescr_frame.refresh(0, 0, self._height - 17, 0, self._height - 8, self._width)
         self._crashescr = self._crashescr_frame.subpad(self._max_log_lines, self._width - 2, 1, 1)
         self._draw_crash()
-        # pytype: enable=module-attr
 
         # Initialise status window
         self._statscr = curses.newwin(6, self._width, self._height - 7, 0)
-        self._statscr.nodelay(1)
+        self._statscr.nodelay(True)
         self._statscr.border()
-        self._statscr.addstr(0, 1, "Status", curses.color_pair(2) | curses.A_BOLD)
+        self._statscr.addstr(0, 1, "Status", curses.color_pair(COLOR_PAIR_CYAN) | curses.A_BOLD)
         self._draw_stat()
 
     def _draw_case(self, indent_size=INDENT_SIZE):
@@ -308,6 +322,7 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
             max_lines=self._max_log_lines,
             total_indent_size=total_indent_size,
             auto_scroll=self._auto_scroll,
+            truncate_long_lines=True,
         )
 
     def _draw_stat(self):
@@ -326,25 +341,25 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
                                                                 self._width - self._indent_size))
         # fmt: on
         # TODO: Get paused flag from sessions
-        if self._status == 0:
+        if self._status == STATUS_RUNNING:
             self._statscr.addstr(4, 1, "Status:")
-            self._statscr.addstr(4, self._indent_size, "Running", curses.color_pair(4))
-        elif self._status == 1:
+            self._statscr.addstr(4, self._indent_size, "Running", curses.color_pair(COLOR_PAIR_YELLOW))
+        elif self._status == STATUS_PAUSED:
             self._statscr.addstr(4, 1, "Status:")
-            self._statscr.addstr(4, self._indent_size, "Paused ", curses.color_pair(3) | curses.A_BLINK)
-        elif self._status == 2:
+            self._statscr.addstr(4, self._indent_size, "Paused ", curses.color_pair(COLOR_PAIR_RED) | curses.A_BLINK)
+        elif self._status == STATUS_DONE:
             self._statscr.addstr(4, 1, "Status:")
-            self._statscr.addstr(4, self._indent_size, "Done   ", curses.color_pair(5))
+            self._statscr.addstr(4, self._indent_size, "Done   ", curses.color_pair(COLOR_PAIR_GREEN))
 
         self._statscr.refresh()
 
     def _draw_screen(self):
         error_counter = 0
         ms_since_refresh = 0
-        k = 0
+        key = 0
         wait_for_key = False
         try:
-            while not ((k == ord("q") or not self._wait_on_quit) and self._quit):
+            while not ((key == ord("q") or not self._wait_on_quit) and self._quit):
                 try:
                     if self._event_resize or ms_since_refresh >= self._refresh_interval:
                         self._draw_main()
@@ -368,7 +383,7 @@ class FuzzLoggerCurses(ifuzz_logger_backend.IFuzzLoggerBackend):
                             self._draw_stat()
                             self._event_case_close = False
 
-                    k = self._stdscr.getch()
+                    key = self._stdscr.getch()
                     curses.flushinp()
 
                     time.sleep(0.1)
@@ -398,7 +413,9 @@ def _progess_bar(current, total, width):
     return title_str + bar_str + percent_str
 
 
-def _render_pad(lines, pad, y_min, x_min, y_max, x_max, max_lines, total_indent_size, auto_scroll):
+def _render_pad(
+    lines, pad, y_min, x_min, y_max, x_max, max_lines, total_indent_size, auto_scroll, truncate_long_lines=False
+):
     total_rows = 0
     height = y_max - y_min + 1
     width = x_max - x_min
@@ -412,25 +429,26 @@ def _render_pad(lines, pad, y_min, x_min, y_max, x_max, max_lines, total_indent_
                 total_rows,
                 0,
                 "Maximum number of lines reached for this window! Increase 'max_log_lines'",
-                curses.color_pair(3),
+                curses.color_pair(COLOR_PAIR_RED),
             )
             total_rows += 1
             break
 
-        columns = width - total_indent_size
-        rows = int(ceil(len(lines[i][0][width:]) / columns))
-        if rows >= 1:
-            for row in range(1, rows + 1):
-                if total_rows < max_lines - 1:
-                    pad.addstr(
-                        total_rows,
-                        total_indent_size,
-                        lines[i][0][width:][(row * columns) - columns : row * columns],
-                        curses.color_pair(lines[i][1]),
-                    )
-                    total_rows += 1
-                else:
-                    break
+        if not truncate_long_lines:
+            columns = width - total_indent_size
+            rows = int(ceil(len(lines[i][0][width:]) / columns))
+            if rows >= 1:
+                for row in range(1, rows + 1):
+                    if total_rows < max_lines - 1:
+                        pad.addstr(
+                            total_rows,
+                            total_indent_size,
+                            lines[i][0][width:][(row * columns) - columns : row * columns],
+                            curses.color_pair(lines[i][1]),
+                        )
+                        total_rows += 1
+                    else:
+                        break
 
     if total_rows > height and auto_scroll:
         offset = total_rows - height